xbackend/requirements.txt
asgiref==3.8.1
Django==5.2
django-extensions==4.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
Pillow==9.5.0
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
